Linear Regression assignment
This code was written as a solution to an assignment to build a multiple linear regression model for the prediction of car prices. You will need to submit a Jupyter notebook for the same. 
Please find below the details of the assignment
 
Problem Statement
A Chinese automobile company Geely Auto aspires to enter the US market by setting up their manufacturing unit there and producing cars locally to give competition to their US and European counterparts. 
They have contracted an automobile consulting company to understand the factors on which the pricing of cars depends. Specifically, they want to understand the factors affecting the pricing of cars in the American market, since those may be very different from the Chinese market. The company wants to know:
Which variables are significant in predicting the price of a car
How well those variables describe the price of a car
Based on various market surveys, the consulting firm has gathered a large dataset of different types of cars across the Americal market. 

Business Goal 
To model the price of cars with the available independent variables. It will be used by the management to understand how exactly the prices vary with the independent variables. They can accordingly manipulate the design of the cars, the business strategy etc. to meet certain price levels. Further, the model will be a good way for management to understand the pricing dynamics of a new market. 

Data Files:
Input Dataset: CarPrice_Assignment.csv
Data Description: Data Dictionary - carprices.xlsx

Solution:
Code for the linear regression model including data preparation - loading, cleaning, transformation and extraction are available in the python notebook "Linear regression assignment_Suleman Safdar.ipynb" 

Model Evaluation & Conclusion:
The r-square of the test model (0.845) is more or less similar to the train model r-square (0.833), therefore we can finalise this model
We can see that the equation of the best fitted line is:
𝑝𝑟𝑖𝑐𝑒=1.188×𝑒𝑛𝑔𝑖𝑛𝑒𝑠𝑖𝑧𝑒+0.1988×𝑏𝑚𝑤+0.2872×𝑟𝑒𝑎𝑟+0.2516×𝑟𝑜𝑡𝑜𝑟−0.0848

Therefore, enginesize is the most important factor to decide the price of the car