Multiple Linear regression - Housing Case study
The python notebook consists of Multiple Linear regression code to optimise the sale prices of the properties based on important factors such as area, bedrooms, parking, etc. 

Problem Statement: A real estate company that has a dataset containing the prices of properties in the Delhi region. It wishes to use the data to essentially :-
1. To identify the variables affecting house prices, e.g. area, number of rooms, bathrooms, etc.
2. To create a linear model that quantitatively relates house prices with variables such as number of rooms, area, number of bathrooms, etc.
3. To know the accuracy of the model, i.e. how well these variables can predict house prices.